Initial Commit
I struggled with the last two bets, but I really enjoyed working on this from the beginning to end. A conversation with Nick Perkins, a technical skills mentor assigned to my group,  really helped  me step back and look at the program in a simpler way. Specifically I was going to try and define the Bins of each wheel in a multidimensional array but instead defined each bin as an object in an object array and each bin had certain properties. Where I stumbled was when I got to the bets. I wasn't sure about what was the most efficient way to pick a random bin representing the ball landing on a bin in the roulette wheel. Also, returning the bins adjacent to the winning bin as you would for split bets, or corner bets, proved to be difficult. I ended up hard-coding enormous switch statements knowing that there was a better way I just have not found yet.